pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opisto-etl"
version = "0.1.0"
description = "ETL to sync seller inventory with Opisto (PostgreSQL -> CSV -> FTPS)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pandas>=2.1,<3",
    "sqlalchemy>=2.0,<3",
    "psycopg2-binary>=2.9,<3",
    "python-dotenv>=1.0,<2",
    "prefect>=3,<4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["opisto_etl*"]

[project.optional-dependencies]
dev = ["pytest"]
